[ADD,DEL] Calculator 클래스 추가, RandomNumberGenerator 클래스 제거, CalculateMatchCountPerRank 기능 추가, CalculateEarningRate 추가
Calculator클래스 로또게임에 필요한 계산을 담당하는 역할입니다.
CalculateMatchCountPerRank 로또 랭크당 유저와 로또일치 개수를 계산하는 메소드 입니다.
CalculateEarningRate는 수익률을 계산하는 기능을 하는 메소드입니다.
